Added profile_body.dart file
