+++ lib/constants/common_vars.dart
@@ -12,3 +12,5 @@ const double common_profile_img_size_s = 30;
 const double common_img_icon_size_s = 16.0;
 
 const double common_icon_width = 44.0;
+
+const double common_radius = 6.0;
